server/requirements.txt
Django==2.0
django-allauth==0.34.0
django-rest-auth==0.9.2
djangorestframework==3.7.3
oauthlib==2.0.6
pytz==2017.3
